Refactor UserGroup view
