--- app/views/UserGroup/AddUserModal/styles.css
@@ -1,9 +1,0 @@
-.modal {
-    display: flex;
-
-    .modal-body {
-        display: flex;
-        align-items: flex-end;
-        gap: var(--dui-spacing-medium);
-    }
-}
