--- app/views/UserGroup/Memberships/styles.css
@@ -1,17 +1,0 @@
-.memberships-container {
-    background-color: var(--dui-color-foreground);
-    padding: 0;
-
-    .header-cell {
-        background-color: var(--dui-color-foreground);
-    }
-
-    .content {
-        padding: 0;
-    }
-
-    .expanded-table {
-        width: 100%;
-        min-height: 40vh;
-    }
-}
