+++ app/views/UserGroup/UserGroupItem/AddUserModal/index.tsx
@@ -19,18 +19,15 @@ import {
 
 import _ts from '#ts';
 import { useLazyRequest } from '#base/utils/restRequest';
-import UserSelectInput from '#components/selections/UserSelectInput';
-import {
-    BasicUser,
-} from '#types';
+import NewUserSelectInput, { User } from '#components/selections/NewUserSelectInput';
 
-import { Membership } from '../AddUsergroupModal';
+import { Membership } from '../../AddUsergroupModal';
 import styles from './styles.css';
 
 type FormType = {
-    member: number;
+    member: string;
     role: string;
-    group?: number;
+    group?: string;
 };
 
 type FormSchema = ObjectSchema<PartialForm<FormType>>;
@@ -65,32 +62,32 @@ const labelSelector = (d: Role) => d.title;
 
 interface Props {
     onModalClose: () => void;
-    group: number;
+    userGroupId: string;
     onUserAddSuccess: () => void;
     userToEdit?: {
-        id: number;
-        member: number;
+        id: string;
+        member: string;
         role: 'admin' | 'normal';
     };
-    users?: BasicUser[];
+    users?: User[];
 }
 
 function AddUserModal(props: Props) {
     const {
         onModalClose,
-        group,
+        userGroupId,
         onUserAddSuccess,
         userToEdit,
         users,
     } = props;
 
-    const [userOptions, setUserOptions] = useState<BasicUser[] | null | undefined>(users);
+    const [userOptions, setUserOptions] = useState<User[] | null | undefined>(users);
 
     const formValue: PartialForm<FormType> = useMemo(() => ({
-        group,
+        group: userGroupId,
         member: userToEdit?.member,
         role: userToEdit?.role,
-    }), [userToEdit, group]);
+    }), [userToEdit, userGroupId]);
 
     const {
         pristine,
@@ -153,7 +150,7 @@ function AddUserModal(props: Props) {
                 </Button>
             )}
         >
-            <UserSelectInput
+            <NewUserSelectInput
                 name="member"
                 value={value.member}
                 onChange={setFieldValue}
@@ -163,12 +160,14 @@ function AddUserModal(props: Props) {
                 disabled={pendingAddMember || isDefined(userToEdit)}
                 label={_ts('usergroup.memberEditModal', 'addMemberLabel')}
                 placeholder={_ts('usergroup.memberEditModal', 'addMemberPlaceholderLabel')}
+                membersExcludeUsergroup={userGroupId}
             />
             <SelectInput
                 name="role"
                 labelSelector={labelSelector}
                 keySelector={keySelector}
                 options={roles}
+                label={_ts('usergroup.memberEditModal', 'roleLabel')}
                 placeholder={_ts('usergroup.memberEditModal', 'roleLabel')}
                 onChange={setFieldValue}
                 value={value.role}
