+++ app/views/UserGroup/styles.css
@@ -11,41 +11,9 @@
         padding: var(--dui-spacing-large);
 
         .user-group-list {
-            min-height: 100%;
-
-            .user-group-item {
-                gap: 0;
-
-                .user-group-header {
-                    border: none;
-                    background-color: var(--dui-color-background);
-
-                    .heading-container {
-                        .heading {
-                            width: auto;
-                            color: var(--dui-color-text);
-                        }
-                    }
-                }
-
-                .heading-description-container {
-                    display: flex;
-                    align-items: center;
-                    gap: var(--dui-spacing-large);
-
-                    .members-value {
-                        order: 1;
-                    }
-                    .members-label {
-                        order: 2;
-                    }
-                }
-
-                .user-group-content {
-                    border: none;
-                    padding: 0;
-                }
-            }
+            display: flex;
+            flex-direction: column;
+            gap: var(--dui-spacing-small);
         }
     }
 }
